Use a random machine identifier by default in GelfMessageChunkEncoder
The previous implementation was using the configured host name of the target server
and would fail if it was shorter than 4 characters. The current implementation
generates a random 4 bytes machine identifier if no other identifier (e. g. the
host name of the client) was provided. Only the first 4 bytes of the provided
machine identifier are being used and the rest is discarded.