Flush queued messages before shutting down (#43)
* Add ability to flush queued messages before shutting down

Some usages of the GELFClient (especially in Lambda functions that execute for a short time) must flush all messages to the network before shutting down.

Without this change, then queued (but unsent) messages will be lost on shutdown. There are valid use cases for both immediately shutting down and for shutting down after flushing messages, so this change adds an additional shutdown method.